+++ SpeechResponder/CustomFunctions/CargoDetails.cs
@@ -16,18 +16,18 @@ namespace EddiSpeechResponder.CustomFunctions
         public string description => Properties.CustomFunctions_Untranslated.CargoDetails;
         public NativeFunction function => new NativeFunction((values) =>
         {
-            var cargo = ConfigService.Instance.cargoMonitorConfiguration?.cargo;
+            var cargoInventory = ConfigService.Instance.cargoMonitorConfiguration?.cargo;
             Cottle.Value value = values[0];
             Cargo result = null;
 
             if (value.Type == Cottle.ValueContent.String)
             {
                 var edname = CommodityDefinition.FromNameOrEDName(value.AsString)?.edname;
-                result = cargo?.FirstOrDefault(c=> c.edname == edname);
+                result = cargoInventory?.FirstOrDefault(c=> c.edname == edname) ?? new Cargo(edname);
             }
             else if (value.Type == Cottle.ValueContent.Number)
             {
-                result = cargo?.FirstOrDefault(c => c.haulageData.FirstOrDefault(h => h.missionid == (long)value.AsNumber) != null);
+                result = cargoInventory?.FirstOrDefault(c => c.haulageData.FirstOrDefault(h => h.missionid == (long)value.AsNumber) != null);
             }
             return new ReflectionValue(result ?? new object());
         }, 1);